Move user.py to repo layer
